luma_demo_app_0/flask_server/requirements.txt
flask==2.3.3
flask-cors==4.0.0
python-dotenv>=1.0.0
requests>=2.32.3
openai>=1.52.0
portia-sdk-python>=0.1.14
anthropic>=0.8.0
pydantic>=2.0.0
typing-extensions>=4.0.0